Prevent __new__ EngineerState to leak into __old__ CoffeeQueue List View.
This part is a bit sensitive because CoffeeQueue List View has references to the EngineerStates in the engineer list view.
So the CoffeeQueue List View must be updated while the Engineer List View still has the __old__ states.
One way to avoid it is to separate the two lists by making each have its own copy of EngineerStates.